requirements.txt
fastapi
uvicorn
requests
pydantic
faiss-cpu
sentence-transformers
matplotlib
